Convert to using Asciidoctor
Asciidoctor PDF provides generation of PDF documents from AsciiDoc
without an intermediate formate.  This greatly improves reliability and
reproducibility.

After coming back to this project I had trouble getting wkhtmltopdf to
produce a reliable result on different machines.  I also did not wish to
go back down the route of using LaTeX as an intermediate format, since
that requires a heavy toolchain and more dependencies.

Asciidoctor is also run with Docker if USE_DOCKER is true (the default).
This makes it so that it's not required to install Ruby and Asciidoctor.