Embed images in HTML version
Embedding images in the HTML version greatly simplifies things.  We no
longer need to ensure the `img/` folder exists in the correct location.
We also no longer need to worry about packaging with `tar` to send a
single file.  The `.html` file can simply be sent.  I leave it up to the
user to compress the `.html` file if necessary.